рефокторинг много

+++ app/src/main/java/com/example/weatherapp/presentation/screen/moon/entity/MoonModel.kt
@@ -2,5 +2,5 @@ package com.example.weatherapp.presentation.screen.moon.entity
 
 data class MoonModel(
     val moonPhase: String,
-    val moonAge: Int
+    val moonAge: Double
 )
